image: Allow `oc image info` to lookup images by wildcard
```
$ oc image info library/mysql:9*
```

now finds matching tags based on the wildcard syntax.
