+++ pkg/cli/image/imagesource/options.go
@@ -6,8 +6,10 @@ import (
 
 	"github.com/docker/distribution"
 	"github.com/openshift/library-go/pkg/image/registryclient"
+	"k8s.io/klog"
 )
 
+// Options contains inputs necessary to build a repository implementation for a reference.
 type Options struct {
 	FileDir             string
 	Insecure            bool
@@ -15,6 +17,7 @@ type Options struct {
 	RegistryContext     *registryclient.Context
 }
 
+// Repository retrieves the appropriate repository implementation for the given typed reference.
 func (o *Options) Repository(ctx context.Context, ref TypedImageReference) (distribution.Repository, error) {
 	switch ref.Type {
 	case DestinationRegistry:
@@ -34,4 +37,34 @@ func (o *Options) Repository(ctx context.Context, ref TypedImageReference) (dist
 	default:
 		return nil, fmt.Errorf("unrecognized image reference type %s", ref.Type)
 	}
+}
+
+// ExpandWildcard expands the provided typed reference (which is known to have an expansion)
+// to a set of explicit image references.
+func (o *Options) ExpandWildcard(ref TypedImageReference) ([]TypedImageReference, error) {
+	reSearch, err := buildTagSearchRegexp(ref.Ref.Tag)
+	if err != nil {
+		return nil, err
+	}
+
+	// lookup tags that match the search
+	repo, err := o.Repository(context.Background(), ref)
+	if err != nil {
+		return nil, err
+	}
+	tags, err := repo.Tags(context.Background()).All(context.Background())
+	if err != nil {
+		return nil, err
+	}
+	klog.V(5).Infof("Search for %q found: %v", ref.Ref.Tag, tags)
+	refs := make([]TypedImageReference, 0, len(tags))
+	for _, tag := range tags {
+		if !reSearch.MatchString(tag) {
+			continue
+		}
+		copied := ref
+		copied.Ref.Tag = tag
+		refs = append(refs, copied)
+	}
+	return refs, nil
 }
