+++ pkg/cli/image/mirror/mappings.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"regexp"
 	"strings"
 	"sync"
 
@@ -257,24 +256,4 @@ func calculateDockerRegistryScopes(tree targetTree) map[contextKey][]auth.Scope 
 		uniqueScopes[registry] = repoScopes
 	}
 	return uniqueScopes
-}
-
-// buildTagSearchRegexp creates a regexp from the provided tag value
-// that can be used to filter tags. It supports standard '*' glob
-// rules.
-func buildTagSearchRegexp(tag string) (*regexp.Regexp, error) {
-	search := tag
-	if (len(search)) == 0 {
-		search = "*"
-	}
-	var reText string
-	for _, part := range strings.Split(search, "*") {
-		if len(part) == 0 {
-			reText += ".*"
-			continue
-		}
-		reText += regexp.QuoteMeta(part)
-	}
-	reText = fmt.Sprintf("^%s$", reText)
-	return regexp.Compile(reText)
 }
